requirements.txt
dependencies for easy setup

flask==2.3.2
opencv-python==4.5.5.64
pygame==2.5.2
twilio==7.16.1
python-dotenv==1.0.0
psutil==5.9.5

Activate the Virtual Environment

cd "A:\vs code programs\Python\Project_folder"
.\env_temp\Scripts\Activate.ps1

Run the Script

& "A:/vs code programs/Python/Project_folder/env_temp/Scripts/python.exe" "a:/vs code programs/Python/Project_folder/03_Face_Recognition.py"

Run the Flask Dashboard
cd "A:\vs code programs\Python\Project_folder"
python app.py
